chore: Refactor unused components
