+++ src/components/common/contentful-content.module.scss
@@ -13,3 +13,27 @@
   padding: spacer(8) 0;
   text-align: center;
 }
+
+.content {
+  h2 {
+    line-height: 1.25;
+
+    @include type-size(500);
+    @media (min-width: $viewport-md) {
+      @include type-size(600);
+    }
+  }
+
+  h2:first-child {
+    margin-top: 0;
+  }
+
+  h3 {
+    @include type-size(400);
+    margin: spacer(32) 0 spacer(4);
+  }
+
+  li {
+    @include margin(16, bottom);
+  }
+}
