+++ src/components/common/infobox.module.scss
@@ -36,10 +36,6 @@
     color: $color-infobox-alert-dark;
   }
 
-  &.question {
-    background-color: $color-infobox-question-light;
-    color: $color-infobox-question-dark;
-  }
   &.full-size {
     height: 100%;
     display: flex;
