--- src/components/common/lists.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import listsStyles from './lists.module.scss'
-
-const OrderedList = ({ children }) => <ol>{children}</ol>
-
-const UnstyledList = ({ children }) => (
-  <ul className={`list-unstyled ${listsStyles.listUnstyled}`}>{children}</ul>
-)
-
-export { OrderedList, UnstyledList }
