--- src/components/common/lists.module.scss
@@ -1,7 +1,0 @@
-.list-unstyled {
-  list-style-type: none;
-  margin: 0;
-  li {
-    margin: 0;
-  }
-}
