--- src/components/common/markdown-content.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import markdownContentStyles from './markdown-content.module.scss'
-
-export default ({ html }) => (
-  <div
-    className={markdownContentStyles.content}
-    dangerouslySetInnerHTML={{ __html: html }}
-  />
-)
