+++ src/components/common/press-list.module.scss
@@ -25,3 +25,9 @@
     opacity: 0.4;
   }
 }
+.source {
+  font-style: italic;
+}
+.title {
+  font-style: normal;
+}
