--- src/components/common/publication.module.scss
@@ -1,6 +1,0 @@
-.source {
-  font-style: italic;
-}
-.title {
-  font-style: normal;
-}
