+++ src/components/pages/state/state-notes.js
@@ -3,7 +3,6 @@ import marked from 'marked'
 import smartypants from 'smartypants'
 
 import Container from '~components/common/container'
-import MarkdownContent from '~components/common/markdown-content'
 import stateNotesStyle from './state-notes.module.scss'
 
 const getBoldedText = text =>
@@ -19,7 +18,10 @@ export default ({ notes, isNarrow = true }) => {
     <Container narrow={isNarrow} className={stateNotesStyle.container}>
       <span className={stateNotesStyle.label}>Notes: </span>
       {notesArray.map(note => (
-        <MarkdownContent key={note} html={smartypants(marked(note))} />
+        <div
+          key={note}
+          dangerouslySetInnerHTML={{ __html: smartypants(marked(note)) }}
+        />
       ))}
     </Container>
   )
