+++ src/pages/data/index.js
@@ -6,7 +6,6 @@ import DetailText from '~components/common/detail-text'
 import Container from '~components/common/container'
 import Layout from '~components/layout'
 import ContentfulContent from '~components/common/contentful-content'
-import MarkdownContent from '~components/common/markdown-content'
 import States from '~components/pages/data/states'
 
 import { DownloadDataRow } from '~components/pages/state/download-data'
@@ -45,8 +44,11 @@ export default ({ data }) => {
       />
       <Container narrow>
         <DetailText>
-          <MarkdownContent
-            html={data.nationalSummaryFootnote.content.childMarkdownRemark.html}
+          <div
+            dangerouslySetInnerHTML={{
+              __html:
+                data.nationalSummaryFootnote.content.childMarkdownRemark.html,
+            }}
           />
         </DetailText>
       </Container>
