+++ src/stories/2-components/12-Infobox.stories.js
@@ -1,9 +1,5 @@
 import React from 'react'
-import {
-  Infobox,
-  QuestionInfobox,
-  AlertInfobox,
-} from '~components/common/infobox'
+import { Infobox, AlertInfobox } from '~components/common/infobox'
 
 export default {
   title: 'Infobox',
@@ -26,10 +22,4 @@ export const alertinfobox = () => (
   <AlertInfobox header="Oh no! This is an alert">
     Something is broken...
   </AlertInfobox>
-)
-
-export const questioninfobox = () => (
-  <QuestionInfobox header="What does this mean?">
-    More information about this question
-  </QuestionInfobox>
 )
